# TestTorneo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Torneos",
        "description": "Documentación de la API para gestionar torneos, jugadores y partidas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/torneos": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Listar todos los torneos",
                "operationId": "4e29c43dc52a51a9209258bf84986dde",
                "responses": {
                    "200": {
                        "description": "Lista de torneos"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Crear un nuevo torneo",
                "operationId": "ce5996daacaeeb4bbe3406485338c5f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "tipo",
                                    "fecha"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Torneo Nacional"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "Masculino",
                                            "Femenino"
                                        ],
                                        "example": "Masculino"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Torneo creado"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/torneos/{id}": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Obtener un torneo por ID",
                "operationId": "d9f455d03020f7d1756fc201a2f5ad88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Torneo encontrado"
                    },
                    "404": {
                        "description": "Torneo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Actualizar un torneo",
                "operationId": "efcf4386945898751d6517127e6b96e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "Masculino",
                                            "Femenino"
                                        ]
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Torneo actualizado"
                    },
                    "404": {
                        "description": "Torneo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Eliminar un torneo",
                "operationId": "2fb17d30739a7af8777c590e66ef1f7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Torneo eliminado"
                    },
                    "404": {
                        "description": "Torneo no encontrado"
                    }
                }
            }
        },
        "/api/torneos/{id}/partidas": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Obtener todas las partidas de un torneo",
                "operationId": "edea75c92b062fd33c5a163c3984f390",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de partidas del torneo"
                    }
                }
            }
        },
        "/api/torneos/{id}/asignar-jugadores": {
            "post": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Asignar jugadores a un torneo",
                "operationId": "0fa41f499be45498e2fe474874d698d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jugadores": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jugadores asignados correctamente"
                    },
                    "400": {
                        "description": "Error al asignar jugadores"
                    }
                }
            }
        },
        "/api/torneos/{id}/comenzar": {
            "post": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Comenzar un torneo",
                "operationId": "ff5294ab3a96fa5951dd22c654540619",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Torneo comenzado"
                    },
                    "400": {
                        "description": "No se pudo comenzar el torneo"
                    }
                }
            }
        },
        "/api/torneos/{id}/estado": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Obtener el estado actual de un torneo",
                "operationId": "5506f916aa367f6ec49106f36f105ca3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del torneo"
                    },
                    "404": {
                        "description": "Torneo no encontrado"
                    }
                }
            }
        },
        "/api/torneos/{id}/partidas/{ronda}": {
            "get": {
                "tags": [
                    "Torneos"
                ],
                "summary": "Obtener partidas de una ronda específica de un torneo",
                "operationId": "14cb71947898f0a5950e5347461961d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ronda",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de partidas de la ronda"
                    },
                    "404": {
                        "description": "No hay partidas en esta ronda"
                    }
                }
            }
        }
    },
...
}